Add Form Builder Template
-Added form builder template to display on the index.html.twig.

-Still in progress with our custom theme.
