+++ config/packages/twig.yaml
@@ -3,3 +3,4 @@ twig:
     debug: '%kernel.debug%'
     strict_variables: '%kernel.debug%'
     exception_controller: null
+    form_themes: ['bootstrap_4_layout.html.twig']
