+++ src/Form/UserProfileType.php
@@ -4,10 +4,12 @@
 
     use Symfony\Component\Form\AbstractType;
     use Symfony\Component\Form\FormBuilderInterface;
+    use Symfony\Component\OptionsResolver\OptionsResolver;
 
     use Symfony\Component\Form\Extension\Core\Type\TextType;
     use Symfony\Component\Form\Extension\Core\Type\EmailType;
     use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+    use Symfony\Component\Form\Extension\Core\Type\FileType;
     use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 
     class UserProfileType extends AbstractType
@@ -16,10 +18,31 @@
         public function buildForm(FormBuilderInterface $builder, array $options) {
 
             $builder
-                ->add('username', TextType::class)
-                ->add('email', EmailType::class)
-                ->add('password', PasswordType::class)
-            ;
+            ->add('username', TextType::class, [
+                'attr' => [
+                        'placeholder' => 'Username'
+                ]
+            ])
+
+            ->add('email', EmailType::class, [
+                'attr' => [
+                        'placeholder' => 'Email address'
+                ]
+            ])
+
+            ->add('password', PasswordType::class, [
+                'attr' => [
+                        'placeholder' => 'Password'
+                ]
+            ])
+
+            ->add('image', FileType::class, [
+                'attr' => [
+                    'placeholder' => 'Upload An Image'
+                ]
+                
+            ]);
+
         }
     }
 ?>
